requirements.txt
langgraph
langchain
langchain-google-genai
langchain-community
faiss-cpu
flask
python-dotenv
typing-extensions
langchain-core
langchain-aws
